refactor: extract file preview into standalone plugins
Split the file preview feature out of Core into two installable plugins:

1. **File Preview** — text, images (PNG/JPG/GIF/BMP/SVG/AVIF), directory
   stats, binary fallback. No heavy dependencies (only optional Pillow
   for AVIF). Self-contained with own key bindings (F3, Shift+F3, Escape,
   Cmd+Y on Mac).

2. **File Preview PDF** — addon that monkey-patches PreviewWidget to add
   PDF rendering via PyMuPDF. Loads after File Preview and extends it.
   Works independently — if PyMuPDF isn't installed, base plugin still
   works without PDF.

Both ship as built-in plugins in Plugins/ but can also be installed
as third-party plugins by copying to DATA_DIRECTORY/Plugins/Third-party/.

Also updated AGENTS.md with plugin system documentation.
