+++ src/main/resources/base/Plugins/Core/Key Bindings (Mac).json
@@ -25,7 +25,5 @@
 	{ "keys": ["Cmd+Up"], "command": "go_up" },
 	{ "keys": ["Shift+Space"], "command": "quick_look" },
 	{ "keys": ["Cmd+Enter"], "command": "open_selected_files" },
-	{ "keys": ["Cmd+Ctrl+F"], "command": "toggle_fullscreen" },
-	{ "keys": ["F3"], "command": "toggle_preview" },
-	{ "keys": ["Cmd+Y"], "command": "toggle_preview" }
+	{ "keys": ["Cmd+Ctrl+F"], "command": "toggle_fullscreen" }
 ]
