+++ src/main/resources/base/Plugins/File Preview/Key Bindings (Mac).json
@@ -1,0 +1,4 @@
+[
+	{ "keys": ["F3"], "command": "toggle_preview" },
+	{ "keys": ["Cmd+Y"], "command": "toggle_preview" }
+]
